use bevy_math for pos and map

+++ classes/src/round_stats.rs
@@ -1,4 +1,4 @@
 #[derive(Default)]
 pub struct RoundStats {
-    pub moves: usize,
+    pub moves: u32,
 }
